Revert "Be explicit about what buildslave to import instead of fishing around."
This change causes Buildbot 0.8 slaves to fail to connect:

Traceback (most recent call last):
  File "run_slave.py", line 471, in <module>
    main()
  File "run_slave.py", line 444, in main
    HotPatchSlaveBuilder(is_testing)
  File "run_slave.py", line 158, in HotPatchSlaveBuilder
    Log('HotPatchSlaveBuilder(%s)' % is_testing)
  File "run_slave.py", line 54, in Log
    from buildbot.slave.bot import log
ImportError: No module named buildbot.slave.bot
make: *** [start] Error 1

This reverts commit 155900.

TBR=maruel@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10919205

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@155947 0039d316-1c4b-4281-b951-d872f2087c98

+++ slave/run_slave.py
@@ -16,9 +16,6 @@ import time
 SCRIPT_PATH = os.path.dirname(os.path.abspath(__file__))
 needs_reboot = False
 
-# Determines which version of buildslave is used. Set by main().
-USE_BUILDBOT_8 = False
-
 # By default, the slave will identify itself to the master by its hostname.
 # To override that, explicitly set a slavename here.
 slavename = None
@@ -49,10 +46,10 @@ def Log(message):
   if not log_imported:
     # pylint: disable=E0611,F0401,W0602
     global log
-    if not USE_BUILDBOT_8:
+    try:
       # buildbot 0.7.12
       from buildbot.slave.bot import log
-    else:
+    except ImportError:
       # buildbot 0.8.x
       from buildslave.bot import log
     log_imported = True
@@ -152,21 +149,18 @@ def Reboot():
   raise Exception('Reboot: Should not return but would have')
 
 
-def HotPatchSlaveBuilder(is_testing):
+def HotPatchSlaveBuilder():
   """We could override the SlaveBuilder class but it's way simpler to just
   hotpatch it."""
-  Log('HotPatchSlaveBuilder(%s)' % is_testing)
   # pylint: disable=E0611,F0401
-  if not USE_BUILDBOT_8:
+  try:
     # buildbot 0.7.12
     from buildbot.slave.bot import Bot, SlaveBuilder
     RemoteShutdownClass = SlaveBuilder
-    Log('Imported 0.7.x')
-  else:
+  except ImportError:
     # buildbot 0.8.x
     from buildslave.bot import Bot, SlaveBuilder
     RemoteShutdownClass = Bot
-    Log('Imported 0.8.x')
   old_remote_shutdown = RemoteShutdownClass.remote_shutdown
 
   def rebooting_remote_shutdown(self):
@@ -185,37 +179,28 @@ def HotPatchSlaveBuilder(is_testing):
     that error is caught.
     """
     global needs_reboot
-    if not is_testing:
-      needs_reboot = True
-      old_remote_shutdown(self)
-    else:
-      Log('Faking Reboot')
+    needs_reboot = True
+    old_remote_shutdown(self)
 
   RemoteShutdownClass.remote_shutdown = rebooting_remote_shutdown
 
   Bot.old_remote_setBuilderList = Bot.remote_setBuilderList
   def cleanup(self, wanted):
     retval = self.old_remote_setBuilderList(wanted)
-    wanted_dirs = sorted(['info', 'cert', '.svn'] + [r[1] for r in wanted])
-    Log('Wanted directories: %s' % wanted_dirs)
-    actual_dirs = sorted(
-        i for i in os.listdir(self.basedir)
-        if os.path.isdir(os.path.join(self.basedir, i)))
-    Log('Actual directories: %s' % actual_dirs)
-    from common import chromium_utils
-    for d in actual_dirs:
+    wanted_dirs = ['info', 'cert', '.svn'] + [r[1] for r in wanted]
+    for d in os.listdir(self.basedir):
       # Delete build.dead directories.
       possible_build_dead = os.path.join(self.basedir, d, 'build.dead')
       if os.path.isdir(possible_build_dead):
+        from common import chromium_utils
         Log('Deleting unwanted directory %s' % possible_build_dead)
-        if not is_testing:
-          chromium_utils.RemoveDirectory(possible_build_dead)
+        chromium_utils.RemoveDirectory(possible_build_dead)
 
       # Delete old slave directories.
-      if d not in wanted_dirs:
+      if d not in wanted_dirs and os.path.isdir(os.path.join(self.basedir, d)):
         Log('Deleting unwanted directory %s' % d)
-        if not is_testing:
-          chromium_utils.RemoveDirectory(os.path.join(self.basedir, d))
+        from common import chromium_utils
+        chromium_utils.RemoveDirectory(os.path.join(self.basedir, d))
     return retval
   Bot.new_remote_setBuilderList = cleanup
   Bot.remote_setBuilderList = Bot.new_remote_setBuilderList
@@ -276,10 +261,10 @@ def error(msg):
 
 def main():
   # Use adhoc argument parsing because of twisted's twisted argument parsing.
-  global USE_BUILDBOT_8
+  use_buildbot_8 = False
   if '--use_buildbot_8' in sys.argv:
     sys.argv.remove('--use_buildbot_8')
-    USE_BUILDBOT_8 = True
+    use_buildbot_8 = True
 
   # Change the current directory to the directory of the script.
   os.chdir(SCRIPT_PATH)
@@ -321,7 +306,7 @@ def main():
   active_master = GetActiveMaster(slave.bootstrap, config_bootstrap,
                                   active_slavename)
 
-  if USE_BUILDBOT_8:
+  if use_buildbot_8:
     bb_ver = 'buildbot_slave_8_4'
     tw_ver = 'twisted_10_2'
   else:
@@ -437,11 +422,12 @@ def main():
     error('Platform %s is not implemented yet' % sys.platform)
 
   # This envrionment is defined only when testing the slave on a dev machine.
-  is_testing = 'TESTING_MASTER' in os.environ
-  if not is_testing:
+  if 'TESTING_MASTER' not in os.environ:
     # Don't overwrite the ~/.subversion/config file when TESTING_MASTER is set.
     FixSubversionConfig()
-  HotPatchSlaveBuilder(is_testing)
+    # Do not reboot the workstation or delete unknown directories in the current
+    # directory.
+    HotPatchSlaveBuilder()
 
   import twisted.scripts.twistd as twistd
   twistd.run()
